synergis-ai/backend/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# AI and ML
langchain>=0.0.300
chromadb>=0.4.0
recbole>=1.1.1
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Database and storage
supabase>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# API and networking
httpx>=0.24.0
requests>=2.31.0
websockets>=11.0.0

# Utils
python-multipart>=0.0.6
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.4.0